migrate unit_test_template to pytest
